requirements-dev.txt
flake8
pyyaml
requests
mysql-connector-python
